Improve delete job functionality
This commit introduces the ability for users to delete jobs from the job list.

-Implemented a delete button for each job item in JobList.vue.
-Added a confirmation dialog before deleting a job.
-Integrated the deleteJob method from the job service to perform the deletion.
-Updated the job service store to include the deleteJob function.
-Added a delete function to job controller and job service to handle the deletion logic.
-Added authorization to the delete function to check if the user owns the job.

+++ job-control/src/business/jobs.service.ts
@@ -25,8 +25,14 @@ export const jobsService = {
         return Job.findByIdAndUpdate(id, data, {new: true});
     },
 
-    async delete(id: string) {
-        return Job.findByIdAndDelete(id);
+    async delete(id: string, email: string) {
+        const job = this.getById(id)
+        if(job.email == email){
+            return Job.findByIdAndDelete(id);
+        }
+        else{
+            throw new Error(`Email: ${email} does not match the job owner's email.`);
+        }
     },
 
     async assignCompany(jobId: string, email: string) {
